+++ Vault/src/utopia/vault/nosql/view/DeprecatableView.scala
@@ -1,0 +1,27 @@
+package utopia.vault.nosql.view
+
+import utopia.vault.nosql.template.Deprecatable
+
+/**
+ * Common trait for views which distinguish between active and deprecated items
+ *
+ * @author Mikko Hilpinen
+ * @since 09.07.2025, v1.22
+ */
+trait DeprecatableView[+Repr] extends FilterableView[Repr]
+{
+	// ABSTRACT -----------------------
+	
+	/**
+	 * @return Model used for interacting with the DB and for building conditions
+	 */
+	protected def model: Deprecatable
+	
+	
+	// COMPUTED ----------------------
+	
+	/**
+	 * @return Access to non-deprecated items
+	 */
+	def active = filter(model.nonDeprecatedCondition)
+}